chore: Clean up conditional compilation using `cfg_aliases` and `moddef`
