+++ Cargo.toml
@@ -26,6 +26,7 @@ bytemuck = { version = "1.22.0", features = [
 ] }
 half = { version = "2.5.0", features = ["bytemuck", "num-traits"] }
 macerator-macros = { version = "0.1.1", path = "crates/macerator-macros" }
+moddef = "0.2"
 num-traits = "0.2.0"
 paste = "1"
 
@@ -35,6 +36,9 @@ half = { version = "2.4", features = ["bytemuck", "num-traits", "rand_distr"] }
 pretty_assertions = "1.4.0"
 rand = { version = "0.9.0" }
 wasm-bindgen-test = "0.3.0"
+
+[build-dependencies]
+cfg_aliases = "0.2"
 
 [target.'cfg(target_arch = "wasm32")'.dev-dependencies]
 getrandom = { version = "0.3.1", features = ["wasm_js"] }
