+++ src/backend/arch/scalar.rs
@@ -1,0 +1,28 @@
+use crate::Simd;
+
+use super::WithSimd;
+
+#[derive(Debug, Clone, Copy)]
+#[non_exhaustive]
+#[repr(u8)]
+pub enum Arch {
+    Scalar,
+}
+
+impl Arch {
+    pub fn new() -> Self {
+        Self::Scalar
+    }
+
+    pub fn dispatch<Op: WithSimd>(self, op: Op) -> Op::Output {
+        match self {
+            Arch::Scalar => <crate::scalar::Fallback as Simd>::vectorize(op),
+        }
+    }
+}
+
+impl Default for Arch {
+    fn default() -> Self {
+        Self::new()
+    }
+}
