+++ src/tests/ord.rs
@@ -51,7 +51,7 @@ macro_rules! testgen_cmp {
                     .zip(rhs.iter())
                     .map(|(a, b)| $ty::$reference(a, b))
                     .collect::<Vec<_>>();
-                #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+                #[cfg(x86)]
                 {
                     use $crate::backend::x86::*;
                     #[cfg(all(feature = "nightly", feature = "fp16"))]
@@ -73,7 +73,7 @@ macro_rules! testgen_cmp {
                         assert_eq!(out_ref, out);
                     }
                 }
-                #[cfg(target_arch = "aarch64")]
+                #[cfg(aarch64)]
                 {
                     use $crate::backend::aarch64::NeonFma;
                     if NeonFma::is_available() {
@@ -81,7 +81,7 @@ macro_rules! testgen_cmp {
                         assert_eq!(out_ref, out);
                     }
                 }
-                #[cfg(target_arch = "wasm32")]
+                #[cfg(wasm32)]
                 {
                     use crate::backend::wasm32::Simd128;
                     let out = Simd128::run_vectorized(|| [<$test_fn _impl>]::<Simd128, $ty>(&lhs, &rhs));
@@ -108,7 +108,7 @@ macro_rules! testgen_min_max {
                     .zip(rhs.iter())
                     .map(|(a, b)| $ty::$reference(*a, *b))
                     .collect::<Vec<_>>();
-                #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+                #[cfg(x86)]
                 {
                     use $crate::backend::x86::*;
                     #[cfg(all(feature = "nightly", feature = "fp16"))]
@@ -130,7 +130,7 @@ macro_rules! testgen_min_max {
                         assert_eq!(out_ref, out);
                     }
                 }
-                #[cfg(target_arch = "aarch64")]
+                #[cfg(aarch64)]
                 {
                     use $crate::backend::aarch64::NeonFma;
                     if NeonFma::is_available() {
@@ -138,7 +138,7 @@ macro_rules! testgen_min_max {
                         assert_eq!(out_ref, out);
                     }
                 }
-                #[cfg(target_arch = "wasm32")]
+                #[cfg(aarch64)]
                 {
                     use crate::backend::wasm32::Simd128;
                     let out = Simd128::run_vectorized(|| [<$test_fn _impl>]::<Simd128, $ty>(&lhs, &rhs));
